WIP: 결재 상세정보 페이지 구현 중
유의할 점
: DTO로 join 결과를 받아오고싶다면, select 컬럼 순서와 DTO 생성자 매개변수 순서가 일치해야 함
새롭게 알게된 점
: controller에서 json 파싱

#55

+++ src/main/java/com/goodee/corpdesk/approval/dto/ReqApprovalDTO.java
@@ -16,7 +16,7 @@ public class ReqApprovalDTO {
 	private Long approvalId;
 	private String username;
 	private Integer departmentId;
-    private Long approvalFormId;
+    private Integer approvalFormId;
     private String approvalContent;
 	private Character status;
     private Character approveYn;
