+++ src/main/java/com/goodee/corpdesk/approval/entity/ApprovalForm.java
@@ -23,7 +23,7 @@ public class ApprovalForm extends BaseEntity {
 	
 	@Id
 //    @GeneratedValue(strategy = GenerationType.IDENTITY) // TODO 서버실행시 기본값 넣기위해 임시로 주석처리
-	private Long approvalFormId;
+	private Integer approvalFormId;
 	
 	@Column(nullable = false)
 	private String formTitle;
